Splitting `pod cache` subcommands in separate files
